pyproject.toml
[project]
name = "sandbox-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Jaouedi Ilyes",email = "ilyes.jaouedi@centralesupelec.fr"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "google-adk (>=1.26.0,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "python-telegram-bot (>=22.6,<23.0)",
    "chromadb (>=1.5.2,<2.0.0)",
    "pypdf (>=6.7.5,<7.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
    "sentence-transformers (>=5.2.3,<6.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "opencv-python (>=4.13.0.92,<5.0.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "google-api-python-client (>=2.191.0,<3.0.0)",
    "google-auth-httplib2 (>=0.3.0,<0.4.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "arxiv (>=2.4.0,<3.0.0)",
    "pycurl (>=7.45.7,<8.0.0)",
    "xmltodict (>=1.0.4,<2.0.0)",
    "notebooklm-py (>=0.3.3,<0.4.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "anthropic[vertex] (>=0.84.0,<0.85.0)",
    "matplotlib (>=3.10.8,<4.0.0)",
    "tqdm (>=4.67.3,<5.0.0)",
    "pandas (>=3.0.1,<4.0.0)",
    "prompt-toolkit (>=3.0.52,<4.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
